crononly enviroment

+++ bootstrap.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 set -e
 SHUTDOWN_WAIT_MINUTES=${1:-"1640"}
 # this job will run every 10 mintes on machine as super user
@@ -7,10 +7,13 @@ REMOTE_CRON_LOCAL=~/remotecron
 REMOTE_CRON_RUNNER=~/remotecronrunner.sh
 LOGGER_FLAG="computing4food"
 
-sudo apt-get install boinc-client boinctui -y
-sudo shutdown -h +${SHUTDOWN_WAIT_MINUTES} &
-boinccmd --project_attach http://www.worldcommunitygrid.org/ 866684_3046dac5b56be2d561d0aad4595508b1
-sudo /etc/init.d/boinc-client start
+if [ $CRON_ONLY != 1 ];then
+	sudo apt-get install boinc-client boinctui -y
+	sudo shutdown -h +${SHUTDOWN_WAIT_MINUTES} &
+	boinccmd --project_attach http://www.worldcommunitygrid.org/ 866684_3046dac5b56be2d561d0aad4595508b1
+	sudo /etc/init.d/boinc-client start
+fi
+
 cat > $REMOTE_CRON_RUNNER <<END
 #/bin/sh
 set -e
@@ -22,4 +25,5 @@ logger -t $LOGGER_FLAG "Run remote job"
 $REMOTE_CRON_LOCAL | logger -t ${LOGGER_FLAG}_remote
 END
 chmod +x $REMOTE_CRON_RUNNER
+
 echo "*/10 * * * * $REMOTE_CRON_RUNNER" | crontab -
